Annotate brokerage and put it under the flags
Sixteen functions. getSite returns SiteSpec, which is what makes everything read off a
site spec checked -- the previous commit is what that found.

get_child_site_spec had three ways to raise on a null column, all of them inside the
loop in __init__ whose except only logs, so one bad queue would have taken every child
queue of every queue with it:

- mincore is optional -- the compatibility check above is the only place the resource
  type requires one -- so an unset mincore is now no lower bound rather than a None
  handed to max(). core_count is already at least 1, so 1 leaves every bound alone.
- minrss and maxrss are both scaled here and CRIC leaves them unset for some queues
  (every other reader spells the test `not in (0, None)`), so a queue missing either
  is skipped with a warning naming it.

SiteSpec.minrss and maxrss are float | None rather than int | None. This function
multiplies them by a core ratio, so the specs it puts in the map have always held
floats; the declaration now says so instead of the arithmetic being rounded to fit it.

resolveNucleus returns str | None through a separate name: the parameter is str and
"NULL" is what it turns into None, which is not something a str parameter can hold.

AtlasBrokerUtils declares its endpoint set as a Collection: the try yields dict_keys
and the except an empty dict, and only the membership test below is used on either.

+++ pandaserver/taskbuffer/SiteSpec.py
@@ -141,8 +141,10 @@ class SiteSpec(object):
     # scope -> endpoints of that scope
     ddm_endpoints_input: dict[str, DdmSpec]
     ddm_endpoints_output: dict[str, DdmSpec]
-    maxrss: int | None
-    minrss: int | None
+    # CRIC gives these as integers, but SiteMapper.get_child_site_spec scales them by a core
+    # ratio when it derives a child queue, so a spec in the map can hold either
+    maxrss: float | None
+    minrss: float | None
     direct_access_lan: bool
     direct_access_wan: bool
     tier: str | None
